indextocomponents
