+++ package.json
@@ -33,6 +33,7 @@
     "@zeit/next-css": "^1.0.1",
     "axios": "^0.24.0",
     "d3": "^5.16.0",
+    "d3-simple-slider": "^1.10.4",
     "date-fns": "^2.25.0",
     "formik": "^2.2.9",
     "history": "^5.0.1",
