+++ src/pages/tools/timeseries-sim/components/SubmitButton.tsx
@@ -1,0 +1,12 @@
+import { Button } from "@material-ui/core";
+
+function SubmitButton(props) {
+  return (
+    <div className={props.className}>
+      <Button variant="outlined" color="primary" onClick={props.onClick}>
+        Submit
+      </Button>
+    </div>
+  );
+}
+export default SubmitButton;
