+++ src/pages/tools/timeseries-sim/variables/variables.js
@@ -1,0 +1,129 @@
+export const covidIndicators = [
+  {
+    label: "Daily Deaths",
+    value: "new_deaths",
+  },
+  {
+    label: "Daily Cases",
+    value: "new_cases",
+  },
+  {
+    label: "Daily Deaths per million",
+    value: "new_deaths_per_million",
+  },
+  {
+    label: "Daily Cases per million",
+    value: "new_cases_per_million",
+  },
+
+  {
+    label: "Cumulative Deaths",
+    value: "total_deaths",
+  },
+  {
+    label: "Cumulative Cases",
+    value: "total_cases",
+  },
+
+  {
+    label: "Cumulative Deaths per million",
+    value: "total_deaths_per_million",
+  },
+  {
+    label: "Cumulative Cases per million",
+    value: "total_cases_per_million",
+  },
+  {
+    label: "Biweekly Cases per million",
+    value: "biweekly_cases_per_million",
+  },
+  {
+    label: "Biweekly Deaths per million",
+    value: "biweekly_deaths_per_million",
+  },
+
+  {
+    label: "Biweekly Cases",
+    value: "biweekly_cases",
+  },
+  {
+    label: "Biweekly Deaths",
+    value: "biweekly_deaths",
+  },
+  {
+    label: "Weekly Cases per million",
+    value: "weekly_cases_per_million",
+  },
+  {
+    label: "Weekly Deaths per million",
+    value: "weekly_deaths_per_million",
+  },
+  {
+    label: "Weekly Cases",
+    value: "weekly_cases",
+  },
+  {
+    label: "Weekly Deaths",
+    value: "weekly_deaths",
+  },
+  {
+    label: "Weekly Cases Rate",
+    value: "weekly_cases_rate",
+  },
+  {
+    label: "Weekly Deaths Rate",
+    value: "weekly_deaths_rate",
+  },
+  {
+    label: "Biweekly Cases Rate",
+    value: "biweekly_cases_rate",
+  },
+  {
+    label: "Biweekly Deaths Rate",
+    value: "biweekly_deaths_rate",
+  },
+];
+
+export const similarityMeasures = [
+  {
+    label: "Euclidean Distance",
+    value: "euclidean",
+  },
+  {
+    label: "Manhattan Distance",
+    value: "manhattan",
+  },
+  {
+    label: "Chebyshev Distance",
+    value: "chebyshev",
+  },
+  {
+    label: "Dynamic Time Warping Distance",
+    value: "dtw",
+  },
+  {
+    label: "Longest Common Subsequence Distance",
+    value: "lcs",
+  },
+];
+
+export const continents = [
+  {
+    value: "Africa",
+  },
+  {
+    value: "Asia",
+  },
+  {
+    value: "Australia",
+  },
+  {
+    value: "Europe",
+  },
+  {
+    value: "North America",
+  },
+  {
+    value: "South America",
+  },
+];
